requirements.txt
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
python-dotenv==1.0.0
jsonschema==4.20.0
beautifulsoup4==4.12.2
pydantic==2.5.2
rich==13.7.0 